Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/script-tag-with-source-data-url-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/script-tag-with-source-data-url.html&q=%3Cscript%20src=%22data:,alert(1)%22' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
